LDEV-4599 Duplicate user data in the tool
User data like name or email seems redundant in tool tables. But
otherwise we would have to hit the core tables every time we want to
fetch it. We stick to redundancy until we have a well working 2nd level
cache and we agree to couple tool and core tables more tightly.